requirements.txt
Pillow==9.2.0
pynput==1.7.6
python-dotenv==0.21.0
pywin32==304
requests==2.26.0
